A Computer Science 3 - 3rd Quarter Project
This game was developed by PSHS-CLC students of Batch 2024 for their 3rd Quarter - Computer Science 3 Project. This game was inspired by the infamous mobile application known as 4 Pics, 1 Word. Furthermore, we would like to acknowledge the works and materials of Shubhang Sati, Keshav Dhandhania, and Anant Jain of Common Lounge for aiding us throughout the process.